build: update babel config for js and react
Updated the babel configuration, including support for es2015+ polyfills and
react. The webpack configuration was also updated for with support for js-type
file processing, a feature which integrates with babel via babel-loader. The set
of browsers that will be supported was also specified through an added
browserslist config file and browerserlist's query syntax.