build: prepare build system
Established the key config files and dependencies for the build system:
using Webpack as the module bundler, Babel to process js-type files, and
PostCSS to process css files. The build system incorporates the basic
elements necessary to support both development and production builds of
a react application. And, it will apply syntax transformations and
polyfills which are required to support a target browser environment.